refactor(codex): split config migration helpers
